backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.0.2
python-dotenv==1.1.1
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic-settings==2.2.1
python-multipart==0.0.20
websockets==11.0.3
email-validator==2.1.0
pdfplumber==0.11.4
python-docx==1.1.2
tiktoken==0.7.0 
openai==1.51.2
google-generativeai==0.7.2